# api/swagger.yaml
swagger: '2.0'
info:
  description: >-
    API to validate South Africa mobile phone numbers campaigns. Allows uploading CSV files listing phone numbers for validation and single-number validation.
  version: '1.0.0'
  title: Phone Validator Service API
  contact:
    name: Ruben Correia
    email: rubencorreia@gmail.com
basePath: /api/v1/phone-numbers
schemes:
  - http
produces:
  - application/json

externalDocs:
  description: README
  url: https://giturl/.../README.md

paths:
  /file:
    post:
      operationId: fileUpload
      summary: Upload CSV file for validation
      description: >-
        Endpoint that receives a CSV file to persist phone numbers (take into account the column id to refer to the uniqueness of those numbers).
      consumes:
        - multipart/form-data
      produces:
        - application/json
      parameters:
        - name: file
          in: formData
          required: true
          type: file
      responses:
        200:
          description: >-
            File was successfully processed and result returned.
          schema:
            $ref: '#/definitions/FileProcess'
        400:
          description: >-
            Invalid request. Complementary informational error message is provided in response body.
        500:
          description: >-
            Internal server error indicating something went wrong while processing file.
  /file/{file-id}:
    get:
      operationId: getFileProcess
      summary: Get file processing result data
      description: >-
        Get file processing result data: detailed number processing results, statistical information.
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/FileIdPathParameter'
      responses:
        200:
          description: >-
            The FileProcess with the given ID
          schema:
            $ref: '#/definitions/FileProcess'
        404:
          description: >-
            File process result not found for provided ID.


  /{phone-number}:
    get:
      operationId: validatePhoneNumber
      summary: Endpoint to test a single phone number.
      description: >-
        Validates if phone number is a valid South Africa mobile number.
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/PhoneNumberPathParameter'
      responses:
        200:
          description: >-
            The ProcessDetail with validation result for provided phone number.
            description: >-
              API to validate South Africa mobile phone numbers campaigns. Allows uploading CSV files listing phone numbers for validation and single-number validation.
            version: '1.0.0'
            title: Phone Validator Service API
            contact:
              name: Ruben Correia
              email: rubencorreia@gmail.com
            basePath: /api/v1/phone-numbers
            schemes:
              - http
            produces:
              - application/json

            externalDocs:
              description: README
              url: https://giturl/.../README.md

            paths:
              /file:
                post:
                  operationId: fileUpload
                  summary: Upload CSV file for validation
                  description: >-
                    Endpoint that receives a CSV file to persist phone numbers (take into account the column id to refer to the uniqueness of those numbers).
                  consumes:
                    - text/csv
                  produces:
                    - application/json
                  parameters:
                    - name: file
                      in: body
                      required: true
                  responses:
                    200:
                      description: >-
                        The Campaign was created successfully
                      schema:
                        $ref: '#/definitions/FileProcess'
                    400:
                      description: >-
                        Invalid request. Complementary informational error message is provided in response body.
                    500:
                      description: >-
                        Internal server error indicating something went wrong while processing file.
              /file/{file-id}:
                get:
                  operationId: getFileProcess
                    summary: Get file processing result data
                    description: >-
                      Get file processing result data: detailed number processing results, statistical information.
                  produces:
                    - application/json
                  parameters:
                    - $ref: '#/parameters/FileIdPathParameter'
                  responses:
                    200:
                      description: >-
                        The FileProcess with the given ID
                      schema:
                        $ref: '#/definitions/FileProcess'
                    404:
                      description: >-
                        File process result not found for provided ID.


              /{phone-number}:
                get:
                  operationId: validatePhoneNumber
                  summary: Endpoint to test a single phone number.
                  description: >-
                    Validates if phone number is a valid South Africa mobile number.
                  produces:
                    - application/json
                  parameters:
                    - $ref: '#/parameters/PhoneNumberPathParameter'
                  responses:
                    200:
                      description: >-
                        The ProcessDetail with validation result for provided phone number.
                      schema:
                        $ref: '#/definitions/ProcessDetail'

            definitions:
              FileProcess:
                type: object
                description: >-
                  The processing result for a CSV file.
                properties:
                  id:
                    type: string
                    readOnly: true
                    description: >-
                      The unique and stable identifier for this File Process result.
                      The format is an eleven characters case sensitive string.
                    example: '5UuAotCNR66'
                  originalFilename:
                    type: string
                    description: >-
                      The original filename.
                    example: 'file.csv'
                  processDetails:
                    type: array
                    items:
                      $ref: '#/definitions/ProcessDetail'
                    description: >-
                      The processing details for each phone number in CSV file.
                  processSummary:
                    $ref: '#/definitions/ProcessSummary'
                    description: >-
                      The file processing statistics.
                  created_at:
                    type: string
                    format: date-time
                    description: >-
                      The date of the processing of CSV file.
                    example: "2019-06-10T18:50:35+00:00"
              ProcessDetail:
                type: object
                description: >-
                  The processing detail for each line in CSV file/phone number.
                properties:
                  id:
                    type: integer
                    readOnly: true
                    description: >-
                      The ID associated with (potential) phone number, as provided in CSV file.
                  originalValue:
                    type: string
                    description: >-
                      The original value in CSV file.
                    example: '27735405794_DELETED_1488987214'
                  processedValue:
                    type: string
                    description: >-
                      The processed value (relevant if original value was fixed).
                    example: '27735405794'
                  status:
                    $ref: '#/definitions/ValidationStatus'
                    description: >-
                      The phone number validation status.
                  error:
                    $ref: '#/definitions/ValidationError'
                    description: >-
                      The phone number validation error detail.
                  created_at:
                    type: string
                    format: date-time
                    description: >-
                      The date of the processing of CSV file.
                    example: "2019-06-10T18:50:35+00:00"
              ValidationError:
                type: object
                description: >-
                  The validation error detail.
                properties:
                  code:
                    type: string
                    description: >-
                      The error code.
                    example: 'INVALID_NUMBER_FORMAT'
                  message:
                    type: string
                    description: >-
                      The error message.
                    example: 'Phone number format provided is invalid'
              ProcessSummary:
                type: object
                description: >-
                  The file processing statistics.
                properties:
                  valid:
                    type: integer
                    description: >-
                      Valid numbers count.
                  fixed:
                    type: integer
                    description: >-
                      Fixed numbers count.
                  invalid:
                    type: integer
                    description: >-
                      Invalid numbers count.
                  created:
                    type: integer
                    description: >-
                      Created numbers count (newly inserted in system).
                  updated:
                    type: integer
                    description: >-
                      Updated numbers count (already inserted in system).
              ValidationStatus:
                type: string
                x-extensible-enum:
                  - VALID
                  - FIXED
                  - INVALID
                description: >-
                  The validation status of an phone number.

            parameters:
              FileIdPathParameter:
                name: file-id
                in: path
                required: true
                type: string
                description: >-
                  The ID of the FileProcess to retrieve.
              PhoneNumberPathParameter:
                name: phone-number
                in: path
                required: true
                type: string
                description: >-
                  The Phone number to be validated.

            $ref: '#/definitions/ProcessDetail'

definitions:
  FileProcess:
    type: object
    description: >-
      The processing result for a CSV file.
    properties:
      id:
        type: string
        readOnly: true
        description: >-
          The unique and stable identifier for this File Process result.
          The format is an eleven characters case sensitive string.
        example: '5UuAotCNR66'
      originalFilename:
        type: string
        description: >-
          The original filename.
        example: 'file.csv'
      processDetails:
        type: array
        items:
          $ref: '#/definitions/ProcessDetail'
        description: >-
          The processing details for each phone number in CSV file.
      processSummary:
        $ref: '#/definitions/ProcessSummary'
        description: >-
          The file processing statistics.
      created_at:
        type: string
        format: date-time
        description: >-
          The date of the processing of CSV file.
        example: "2019-06-10T18:50:35+00:00"
  ProcessDetail:
    type: object
    description: >-
      The processing detail for each line in CSV file/phone number.
    properties:
      id:
        type: integer
        readOnly: true
        description: >-
          The ID associated with (potential) phone number, as provided in CSV file.
      originalValue:
        type: string
        description: >-
          The original value in CSV file.
        example: '27735405794_DELETED_1488987214'
      processedValue:
        type: string
        description: >-
          The processed value (relevant if original value was fixed).
        example: '27735405794'
      status:
        $ref: '#/definitions/ValidationStatus'
        description: >-
          The phone number validation status.
      error:
        $ref: '#/definitions/ValidationError'
        description: >-
          The phone number validation error detail.
      created_at:
        type: string
        format: date-time
        description: >-
          The date of the processing of CSV file.
        example: "2019-06-10T18:50:35+00:00"
  ValidationError:
    type: object
    description: >-
      The validation error detail.
    properties:
      code:
        type: string
        description: >-
          The error code.
        example: 'INVALID_NUMBER_FORMAT'
      message:
        type: string
        description: >-
          The error message.
        example: 'Phone number format provided is invalid'
  ProcessSummary:
    type: object
    description: >-
      The file processing statistics.
    properties:
      valid:
        type: integer
        description: >-
          Valid numbers count.
      fixed:
        type: integer
        description: >-
          Fixed numbers count.
      invalid:
        type: integer
        description: >-
          Invalid numbers count.
      created:
        type: integer
        description: >-
          Created numbers count (newly inserted in system).
      updated:
        type: integer
        description: >-
          Updated numbers count (already inserted in system).
  ValidationStatus:
    type: string
    x-extensible-enum:
      - VALID
      - FIXED
      - INVALID
    description: >-
      The validation status of an phone number.

parameters:
  FileIdPathParameter:
    name: file-id
    in: path
    required: true
    type: string
    description: >-
      The ID of the FileProcess to retrieve.
  PhoneNumberPathParameter:
    name: phone-number
    in: path
    required: true
    type: string
    description: >-
      The Phone number to be validated.